requirements.txt
python-mpv>=1.0.4
PyQt6>=6.5.0
darkdetect>=0.8.0
qtawesome>=1.3.0
